+++ foo_spider_monkey_panel/js_engine/js_compartment_inner.h
@@ -21,8 +21,8 @@ public:
     void OnGcDone();
     bool IsMarkedForGc() const;
 
-    uint32_t GetCurrentHeapBytes() const;
-    uint32_t GetLastHeapBytes() const;
+    uint64_t GetCurrentHeapBytes() const;
+    uint64_t GetLastHeapBytes() const;
 
     uint32_t GetCurrentAllocCount() const;
     uint32_t GetLastAllocCount() const;
@@ -34,8 +34,8 @@ private:
     bool isMarkedForDeletion_ = false;
     bool isMarkedForGc_ = false;
     mutable std::mutex gcDataLock_;
-    uint32_t curHeapSize_ = 0;
-    uint32_t lastHeapSize_ = 0;
+    uint64_t curHeapSize_ = 0;
+    uint64_t lastHeapSize_ = 0;
     uint32_t curAllocCount_ = 0;
     uint32_t lastAllocCount_ = 0;
 };
